fix(challenge-record): 인증 등록 성공 후 새로 인증기록 데이터 페칭
- getChallengeRecord(기록 가져오는) 호출 위치를 ChallengeRecord로 변경
- Verification에서 인증 등록 성공하면 activeTab을 변경하고, activeTab이 업데이트되면 getChallengeRecord를 새로 호출함
