+++ src/pages/challenge-record/index.tsx
@@ -1,8 +1,10 @@
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useSearchParams } from 'react-router-dom';
 
 import Records from './records';
 import Verification from './verification';
+import { getChallengeRecord } from '@/apis/challenge-record/challenge.record.api';
+import { ChallengeRecordData } from '@/apis/challenge-record/challenge.record.response';
 import ChallengeTitle from '@/components/common/challenge-title';
 import { Tab, Tabs } from '@/components/common/tabs';
 import { TabPanel, TabPanels } from '@/components/common/tabs/tab-panels';
@@ -23,21 +25,27 @@ const ChallengeRecord = () => {
     return savedTab ? Number(savedTab) : 0;
   });
 
-  // Records -> Verification으로 비동기 데이터 전달
+  const [records, setRecords] = useState<ChallengeRecordData | null>(); // api 응답 데이터 전체
+  const [recordIdList, setRecordIdList] = useState<number[]>([]); // 한 챌린지의 인증기록 id 리스트
+
+  // recordId를 recordList에 추가하는 함수
+  const fillRecordList = (length: number, values: number[]) => {
+    const newRecordIdList = new Array(length).fill(-1); // 모두 -1로 초기화
+    for (let i = 0; i < values.length; i++) {
+      newRecordIdList[i] = values[i]; // recordId로 바꾸기
+    }
+    setRecordIdList(newRecordIdList);
+  };
+
+  // Verification으로 비동기 데이터 전달
   const [successCount, setSuccessCount] = useState<number>(0);
   const [totalCount, setTotalCount] = useState<number>(0);
   const [endDate, setEndDate] = useState<string>('');
-
   const tabsList = [
     {
       label: '인증 기록',
-      panel: (
-        <Records
-          challengeId={challengeId}
-          setSuccessCount={setSuccessCount}
-          setTotalCount={setTotalCount}
-          setEndDate={setEndDate}
-        />
+      panel: records && (
+        <Records records={records} recordIdList={recordIdList} />
       ),
     },
     {
@@ -58,6 +66,29 @@ const ChallengeRecord = () => {
     setActiveTab(value as 0 | 1);
     sessionStorage.setItem('activeTab', String(value));
   };
+
+  // activeTab이 변경될 때 (인증 등록 후) records 데이터를 새로 페칭
+  useEffect(() => {
+    if (activeTab === 0) {
+      getChallengeRecord(challengeId)
+        .then((res) => {
+          setRecords(res);
+          fillRecordList(res.totalCount, res.recordIds);
+          setSuccessCount(res.successCount);
+          setTotalCount(res.totalCount);
+          setEndDate(res.endDate);
+        })
+        .catch((error) => {
+          // API에서 받은 오류 객체일 경우
+          if (error?.result === 'FAIL') {
+            console.error('Error fetching records:', error);
+          } else {
+            // 예상치 못한 오류 처리
+            console.error('Error fetching records: unexpected, ', error);
+          }
+        });
+    }
+  }, [activeTab, challengeId]);
 
   return (
     <>
